[Feat] Create team header
