+++ src/types/inputtypes.tsx
@@ -22,24 +22,3 @@ export type ReplyItemProps = {
   name?: string;
   date?: string;
 };
-
-export interface BaseProfileProps {
-  imageUrl: string;
-  width: number;
-  showEditIcon?: boolean;
-}
-
-export type ProfileProps = {
-  profileUrl?: string;
-  width: number;
-};
-
-export type TeamProfileProps = {
-  teamProfileUrl?: string;
-  width: number;
-};
-
-export interface TaskListTapProps {
-  title: string;
-  isSelected?: boolean;
-}
