+++ src/types/tasktypes.tsx
@@ -1,0 +1,8 @@
+export interface TaskListTapProps {
+  title: string;
+  isSelected?: boolean;
+}
+
+export type TeamHeaderProps = {
+  title: string;
+};
